requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3
maritalk==0.2.6
python-dotenv==1.0.1
pytest==8.3.3



